Now data will be persisted to firestore network db and user can sync/refresh the data from network.

+++ app/src/main/java/com/incture/mobility/architecturecomponents/activities/CreateNoteActivity.java
@@ -27,6 +27,7 @@ import com.incture.mobility.architecturecomponents.R;
 import com.incture.mobility.architecturecomponents.Utils.BaseActivity;
 import com.incture.mobility.architecturecomponents.Utils.Constants;
 import com.incture.mobility.architecturecomponents.Utils.FileUtility;
+import com.incture.mobility.architecturecomponents.Utils.NetworkUtility;
 import com.incture.mobility.architecturecomponents.dagger.ApplicationComponent;
 import com.incture.mobility.architecturecomponents.room.Notes;
 import com.incture.mobility.architecturecomponents.viewmodels.CreateNotesViewModel;
@@ -42,6 +43,7 @@ import javax.inject.Inject;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import dagger.android.AndroidInjection;
 
 /**
  * Created by satiswardash on 11/02/18.
@@ -170,8 +172,12 @@ public class CreateNoteActivity extends BaseActivity {
         mOptionButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                mCreateNotesViewModel.deleteNote(notes);
-                onBackPressed();
+                if (NetworkUtility.hasNetworkAccess(getApplicationContext())) {
+                    mCreateNotesViewModel.deleteNote(notes);
+                    onBackPressed();
+                }
+                else
+                    Toast.makeText(CreateNoteActivity.this, "You cannot delete data while offline. ", Toast.LENGTH_SHORT).show();
             }
         });
